# server/openapi.yaml
openapi: 3.1.0
info:
  version: 0.1.0
  title: epghub API
  description: 録画管理 API (epghub). 録画特化・TVDB 連携・視聴機能なし。 スキーマ駆動設計で OpenAPI は zod から自動生成される。
servers:
  - url: http://localhost:3000
    description: dev
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: reserve.conflict
        message:
          type: string
          example: 予約時間帯に空きチューナーがありません
        detail:
          type: object
          additionalProperties: {}
      required:
        - code
        - message
...
    RecordedHistory:
      type: object
      properties:
        id:
          type: integer
        tvdbId:
          type:
            - integer
            - "null"
        season:
          type:
            - integer
            - "null"
        episode:
          type:
            - integer
            - "null"
        normalizedTitle:
          type:
            - string
            - "null"
        endAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tvdbId
        - season
        - episode
        - normalizedTitle
        - endAt
        - createdAt
...
    DiskSweepResult:
      type: object
      properties:
        deletedIds:
          type: array
          items:
            type: string
          description: 削除した recordings.id の一覧 (古い順)
        freedBytes:
          type: integer
          description: 削除で解放した推定バイト数 (recordings.size * GiB)
      required:
        - deletedIds
        - freedBytes
    DiskSweepBody:
      type: object
      properties:
        minFreeBytes:
          type: integer
          exclusiveMinimum: 0
          description: 閾値のワンショット上書き (bytes)。通常は env DISK_SWEEP_MIN_FREE_GB または総容量の 5%
            を使うが、強制 sweep や e2e テストで任意の閾値を指定したい場合に送る。
...
    ChannelSourceKind:
      type: string
      enum:
        - mirakurun
        - iptv
      description: デバイス種別
    ChannelSource:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: IPTV JP
        kind:
          $ref: "#/components/schemas/ChannelSourceKind"
        url:
          type: string
          format: uri
          example: http://mirakurun:40772/api/iptv
        xmltvUrl:
          type:
            - string
            - "null"
          format: uri
          description: iptv kind のみ。番組表 XMLTV フィードの URL (任意)
          example: http://mirakurun:40772/api/iptv/xmltv
        friendlyName:
          type:
            - string
            - "null"
          example: Mirakurun
        model:
          type:
            - string
            - "null"
          example: Mirakurun
        deviceId:
          type:
            - string
            - "null"
          example: A1B2C3D4
        tunerCount:
          type:
            - integer
            - "null"
          example: 8
        enabled:
          type: boolean
          example: true
        lastSyncAt:
          type:
            - string
            - "null"
          format: date-time
        lastError:
          type:
            - string
            - "null"
        channelCount:
          type: integer
          description: 前回の sync で登録できたチャンネル件数
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - kind
        - url
        - xmltvUrl
        - friendlyName
        - model
        - deviceId
        - tunerCount
        - enabled
        - lastSyncAt
        - lastError
        - channelCount
        - createdAt
    ChannelSourceList:
      type: array
      items:
        $ref: "#/components/schemas/ChannelSource"
    CreateChannelSource:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: IPTV JP
        kind:
          $ref: "#/components/schemas/ChannelSourceKind"
        url:
          type: string
          format: uri
          example: http://mirakurun:40772/api/iptv
        xmltvUrl:
          type:
            - string
            - "null"
          format: uri
          description: iptv kind のみ。番組表 XMLTV フィードの URL
          example: http://mirakurun:40772/api/iptv/xmltv
      required:
        - name
        - kind
        - url
...
    ProbeChannelSourceResult:
      type: object
      properties:
        reachable:
          type: boolean
        friendlyName:
          type:
            - string
            - "null"
        model:
          type:
            - string
            - "null"
        tunerCount:
          type:
            - integer
            - "null"
        suggestedXmltvUrl:
          type:
            - string
            - "null"
        inferredKind:
          type:
            - string
            - "null"
      required:
        - reachable
        - friendlyName
        - model
        - tunerCount
        - suggestedXmltvUrl
        - inferredKind
    ProbeChannelSource:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: http://mirakurun:40772/api/iptv
      required:
        - url
    ChannelSourceSyncResult:
      type: object
      properties:
        channelCount:
          type: integer
        error:
          type: string
      required:
        - channelCount
    GpuEncoder:
      type: string
      enum:
        - h264_nvenc
        - hevc_nvenc
        - h264_vaapi
        - hevc_vaapi
        - h264_qsv
        - hevc_qsv
        - h264_videotoolbox
        - hevc_videotoolbox
    GpuProbeDetail:
      type: object
      properties:
        ok:
          type: boolean
          description: この encoder が 0.5s テスト encode に成功したか
        error:
          type: string
          description: 失敗時の短い ffmpeg エラー (stderr 末尾 3 行まで)
      required:
        - ok
    GpuProbeResult:
      type: object
      properties:
        available:
          type: array
          items:
            $ref: "#/components/schemas/GpuEncoder"
          description: ok=true だった encoder の一覧 (details から射影)
        details:
          type: object
          properties:
            h264_nvenc: &a6
              $ref: "#/components/schemas/GpuProbeDetail"
            hevc_nvenc: *a6
            h264_vaapi: *a6
            hevc_vaapi: *a6
            h264_qsv: *a6
            hevc_qsv: *a6
            h264_videotoolbox: *a6
            hevc_videotoolbox: *a6
          description: 8 encoder 全ての試行結果 (未検出でも key は必ず存在する)
        probedAt:
          type: string
          format: date-time
          description: probe を走らせた時刻 (ISO-8601)
      required:
        - available
        - details
        - probedAt
    GpuStatus:
      type: object
      properties:
        enabled:
          type: boolean
          description: 録画エンコード時に GPU プリセットへ自動昇格するか
        preferred:
          allOf:
            - $ref: "#/components/schemas/GpuEncoder"
            - type:
                - string
                - "null"
              description: enabled=true 時に使う encoder。null なら resolvePreset は何もしない
        lastProbe:
          allOf:
            - $ref: "#/components/schemas/GpuProbeResult"
            - type:
                - object
                - "null"
              description: 直近の probe 結果。未実行なら null
      required:
        - enabled
        - preferred
        - lastProbe
    GpuSettingsPatch:
      type: object
      properties:
        enabled:
          type: boolean
        preferred:
          allOf:
            - $ref: "#/components/schemas/GpuEncoder"
            - type:
                - string
                - "null"
...
paths:
...
  /admin/recorded-history:
    get:
      tags:
        - admin
      summary: 録画履歴を一覧
      parameters:
        - schema:
            type:
              - integer
              - "null"
            description: 指定時はこの TVDB ID に紐づく行のみ返す
          required: false
          name: tvdbId
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: 上限件数 (デフォルト 200)
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: 履歴行 (createdAt 降順)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecordedHistory"
...
  /admin/capacity/sweep:
    post:
      tags:
        - admin
      summary: ディスク sweep を今すぐ実行
      description: 通常 pg-boss が 1 時間ごとに実行している capacityService.sweep() を同期的に走らせる。保護
        (protected=true) されていない state=ready/failed の行を古い順に削除し、空き容量が
        閾値を上回ったら停止する。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiskSweepBody"
      responses:
        "200":
          description: 削除結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiskSweepResult"
...
  /admin/channel-sources:
    get:
      tags:
        - admin
      summary: チャンネルソース一覧
      description: 登録済みの m3u / Mirakurun アップストリームを取得する。
      responses:
        "200":
          description: 登録済みソース (新しい順)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSourceList"
    post:
      tags:
        - admin
      summary: チャンネルソースを追加
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelSource"
      responses:
        "201":
          description: 追加した行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSource"
        "400":
          description: 入力エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/channel-sources/probe:
    post:
      tags:
        - admin
      summary: チャンネルソース URL の自動検出
      description: HDHomeRun /discover.json を叩いてデバイス情報を取得し、既知の URL パターン (Mirakurun /
        EPGStation) から XMLTV URL を推測する。追加モーダルの自動入力用。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbeChannelSource"
      responses:
        "200":
          description: probe 結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbeChannelSourceResult"
  /admin/channel-sources/{id}/sync:
    post:
      tags:
        - admin
      summary: チャンネルソースを今すぐ同期
      parameters:
        - schema:
            type:
              - integer
              - "null"
          required: false
          name: id
          in: path
      responses:
        "200":
          description: sync 結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSourceSyncResult"
        "404":
          description: 存在しない id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/channel-sources/{id}:
    delete:
      tags:
        - admin
      summary: チャンネルソースを削除
      parameters:
        - schema:
            type:
              - integer
              - "null"
          required: false
          name: id
          in: path
      responses:
        "204":
          description: 削除完了
        "404":
          description: 存在しない id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/gpu/settings:
    patch:
      tags:
        - admin
      summary: GPU エンコード設定を更新
      description: enabled/preferred を部分更新する。送らなかった key は変更しない。 preferred を null にすると
        resolvePreset は CPU プリセットにフォールバックする。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GpuSettingsPatch"
      responses:
        "200":
          description: 更新後のステータス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GpuStatus"
...